requirements.txt
PySide6
qdarkstyle
mido
PyYAML
